+++ packages/contact-center/store/tests/storeEventsWrapper.ts
@@ -12,7 +12,26 @@ jest.mock('../src/store', () => ({
       off: jest.fn(),
     },
     logger: 'mockLogger',
-    idleCodes: 'mockIdleCodes',
+    idleCodes: [
+      {
+        id: 'mockId1',
+        name: 'mockName',
+        isSystem: false,
+        isDefault: false,
+      },
+      {
+        id: 'mockId2',
+        name: 'RONA',
+        isSystem: true,
+        isDefault: false,
+      },
+      {
+        id: 'mockId3',
+        name: 'RONA2',
+        isSystem: true,
+        isDefault: false,
+      },
+    ],
     agentId: 'mockAgentId',
     wrapupCodes: 'mockWrapupCodes',
     currentTask: 'mockCurrentTask',
@@ -25,9 +44,11 @@ jest.mock('../src/store', () => ({
     lastStateChangeTimestamp: 'mockLastStateChangeTimestamp',
     showMultipleLoginAlert: 'mockShowMultipleLoginAlert',
     currentTheme: 'mockCurrentTheme',
+    customState: 'mockCustomState',
     setShowMultipleLoginAlert: jest.fn(),
     setCurrentState: jest.fn(),
     setLastStateChangeTimestamp: jest.fn(),
+    setLastIdleCodeChangeTimestamp: jest.fn(),
     setDeviceType: jest.fn(),
     init: jest.fn().mockResolvedValue({}),
     setIncomingTask: jest.fn(),
@@ -54,8 +75,9 @@ describe('storeEventsWrapper', () => {
       expect(storeWrapper.logger).toBe('mockLogger');
     });
 
-    it('should proxy idleCodes', () => {
-      expect(storeWrapper.idleCodes).toBe('mockIdleCodes');
+    it('should proxy idleCodes and include RONA', () => {
+      expect(storeWrapper.idleCodes.length).toBe(2);
+      expect(storeWrapper.idleCodes[1].name).toBe('RONA');
     });
 
     it('should proxy agentId', () => {
@@ -98,6 +120,10 @@ describe('storeEventsWrapper', () => {
       expect(storeWrapper.currentState).toBe('mockCurrentState');
     });
 
+    it('should proxy customState', () => {
+      expect(storeWrapper.customState).toBe('mockCustomState');
+    });
+
     it('should proxy lastStateChangeTimestamp', () => {
       expect(storeWrapper.lastStateChangeTimestamp).toBe('mockLastStateChangeTimestamp');
     });
@@ -123,6 +149,31 @@ describe('storeEventsWrapper', () => {
       expect(storeWrapper['store'].currentState).toBe('newState');
     });
 
+    describe('setState', () => {
+      it('should call setCurrentState if idleCode is passed', () => {
+        const idleCode = storeWrapper.idleCodes[0];
+        storeWrapper.setState(idleCode);
+        expect(storeWrapper.currentState).toBe(idleCode.id);
+      });
+
+      it('should set customState if customState is passed', () => {
+        const customState = {
+          name: 'customState',
+          developerName: 'customState',
+        };
+        storeWrapper.setState(customState);
+        expect(storeWrapper.customState).toBe(customState);
+      });
+
+      it('should set customState to null if reset is passed', () => {
+        const customState = {
+          reset: true,
+        };
+        storeWrapper.setState(customState);
+        expect(storeWrapper.customState).toBe(null);
+      });
+    });
+
     it('should call registerCC', () => {
       const mockRegisterCC = jest.fn();
       storeWrapper['store'].registerCC = mockRegisterCC;
@@ -144,6 +195,14 @@ describe('storeEventsWrapper', () => {
       const timestamp = new Date();
       storeWrapper.setLastStateChangeTimestamp(timestamp);
       expect(storeWrapper['store'].lastStateChangeTimestamp).toBe(timestamp);
+    });
+
+    it('should setLastIdleCodeChangeTimestamp', () => {
+      expect(storeWrapper.setLastIdleCodeChangeTimestamp).toBeInstanceOf(Function);
+
+      const timestamp = new Date();
+      storeWrapper.setLastIdleCodeChangeTimestamp(timestamp);
+      expect(storeWrapper['store'].lastIdleCodeChangeTimestamp).toBe(timestamp);
     });
 
     it('should currentTheme', () => {