+++ packages/contact-center/user-state/tests/helper.ts
@@ -15,7 +15,12 @@ describe('useUserState Hook', () => {
   ];
 
   const agentId = 'agent123';
-  let workerMock;
+  let workerMock, blobMock;
+  const onStateChange = jest.fn();
+  const logger = {
+    log: jest.fn(),
+    error: jest.fn(),
+  };
 
   beforeEach(() => {
     jest.useFakeTimers();
@@ -30,9 +35,11 @@ describe('useUserState Hook', () => {
     };
 
     global.Worker = jest.fn(() => workerMock);
-    global.URL.createObjectURL = jest.fn(() => 'blob:http://localhost:3000/12345');
+    blobMock = jest.fn(() => 'blob:http://localhost:3000/12345');
+    global.URL.createObjectURL = blobMock;
     jest.spyOn(store, 'setCurrentState');
     jest.spyOn(store, 'setLastStateChangeTimestamp');
+    jest.spyOn(console, 'error').mockImplementation(() => {});
   });
 
   afterEach(() => {
@@ -42,75 +49,333 @@ describe('useUserState Hook', () => {
 
   it('should initialize with default values', () => {
     const {result} = renderHook(() =>
-      useUserState({idleCodes, agentId, cc: mockCC, currentState: '0', lastStateChangeTimestamp: new Date()})
+      useUserState({
+        idleCodes,
+        agentId,
+        cc: mockCC,
+        currentState: '0',
+        customState: null,
+        logger,
+        onStateChange,
+        lastStateChangeTimestamp: new Date().getTime(),
+        lastIdleCodeChangeTimestamp: undefined,
+      })
     );
 
     expect(result.current).toMatchObject({
       isSettingAgentStatus: false,
-      errorMessage: '',
       elapsedTime: 0,
       currentState: '0',
     });
   });
 
-  it('should increment elapsedTime every second', () => {
-    const {result} = renderHook(() =>
-      useUserState({idleCodes, agentId, cc: mockCC, currentState: '0', lastStateChangeTimestamp: undefined})
+  it('should clean up on unmount', () => {
+    mockCC.setAgentState.mockResolvedValueOnce({
+      data: {auxCodeId: '2', lastStateChangeTimestamp: new Date().getTime()},
+    });
+    const {unmount} = renderHook(() =>
+      useUserState({
+        idleCodes: [],
+        agentId: 'agent123',
+        cc: {},
+        currentState: '0',
+        customState: null,
+        logger: {log: jest.fn(), error: jest.fn()},
+        onStateChange: jest.fn(),
+        lastStateChangeTimestamp: new Date().getTime(),
+        lastIdleCodeChangeTimestamp: undefined,
+      })
     );
 
-    act(() => {
-      workerMock.onmessage({data: 1});
-      jest.advanceTimersByTime(1000);
-      workerMock.onmessage({data: 2});
-      jest.advanceTimersByTime(1000);
-      workerMock.onmessage({data: 3});
-    });
+    // Simulate component unmount
+    unmount();
 
-    expect(result.current.elapsedTime).toBe(3);
+    expect(workerMock.postMessage).toHaveBeenCalledWith({type: 'stop'});
+    expect(workerMock.postMessage).toHaveBeenCalledWith({type: 'stopIdleCode'});
+    expect(workerMock.terminate).toHaveBeenCalled();
   });
 
-  it('should handle setAgentStatus correctly and update state', async () => {
-    mockCC.setAgentState.mockResolvedValueOnce({data: {auxCodeId: '2', lastStateChangeTimestamp: new Date()}});
+  it('should increment elapsedTime every second', async () => {
     const {result} = renderHook(() =>
-      useUserState({idleCodes, agentId, cc: mockCC, currentState: '0', lastStateChangeTimestamp: new Date()})
+      useUserState({
+        idleCodes,
+        agentId,
+        cc: mockCC,
+        currentState: '0',
+        customState: null,
+        logger,
+        onStateChange,
+        lastStateChangeTimestamp: new Date().getTime(),
+        lastIdleCodeChangeTimestamp: new Date().getTime(),
+      })
     );
 
     act(() => {
-      result.current.setAgentStatus(idleCodes[1]);
+      workerMock.onmessage({data: {type: 'elapsedTime', elapsedTime: 1}});
+      jest.advanceTimersByTime(1000);
+      workerMock.onmessage({data: {type: 'elapsedTime', elapsedTime: 2}});
+      jest.advanceTimersByTime(1000);
+      workerMock.onmessage({data: {type: 'elapsedTime', elapsedTime: 3}});
     });
 
     await waitFor(() => {
-      expect(store.setCurrentState).toHaveBeenCalledWith('2');
+      expect(result.current.elapsedTime).toBe(3);
     });
   });
 
-  it('should handle setAgentStatus correctly and update state', async () => {
-    mockCC.setAgentState.mockResolvedValueOnce({data: {auxCodeId: '2', lastStateChangeTimestamp: new Date()}});
+  it('should increment lastIdleStateChangeElapsedTime every second', () => {
     const {result} = renderHook(() =>
-      useUserState({idleCodes, agentId, cc: mockCC, currentState: '0', lastStateChangeTimestamp: new Date()})
+      useUserState({
+        idleCodes,
+        agentId,
+        cc: mockCC,
+        currentState: '0',
+        customState: null,
+        logger,
+        onStateChange,
+        lastStateChangeTimestamp: new Date().getTime(),
+        lastIdleCodeChangeTimestamp: new Date().getTime(),
+      })
     );
 
     act(() => {
-      result.current.setAgentStatus(idleCodes[0]);
+      workerMock.onmessage({data: {type: 'lastIdleStateChangeElapsedTime', elapsedTime: 1}});
+      jest.advanceTimersByTime(1000);
+      workerMock.onmessage({data: {type: 'lastIdleStateChangeElapsedTime', elapsedTime: 2}});
+      jest.advanceTimersByTime(1000);
+      workerMock.onmessage({data: {type: 'lastIdleStateChangeElapsedTime', elapsedTime: 3}});
+    });
+
+    expect(result.current.lastIdleStateChangeElapsedTime).toBe(3);
+  });
+
+  it('should handle setAgentStatus correctly and update state', async () => {
+    mockCC.setAgentState.mockResolvedValueOnce({
+      data: {auxCodeId: '2', lastStateChangeTimestamp: new Date().getTime()},
+    });
+    const {result} = renderHook(() =>
+      useUserState({
+        idleCodes,
+        agentId,
+        cc: mockCC,
+        currentState: '0',
+        customState: null,
+        logger,
+        onStateChange,
+        lastStateChangeTimestamp: new Date().getTime(),
+        lastIdleCodeChangeTimestamp: undefined,
+      })
+    );
+
+    act(() => {
+      result.current.setAgentStatus(idleCodes[1].id);
     });
 
     await waitFor(() => {
-      expect(store.setCurrentState).toHaveBeenCalledWith('2');
+      expect(store.setCurrentState).toHaveBeenCalledWith(idleCodes[1].id);
     });
   });
 
-  it('should handle errors from setAgentStatus and revert state', async () => {
+  it('should update last state change timestamp from setAgentState', async () => {
+    const resolvingValue = {data: {auxCodeId: '2', lastStateChangeTimestamp: 1740748111287}};
+    mockCC.setAgentState.mockResolvedValueOnce(resolvingValue);
+    const {rerender} = renderHook(
+      ({currentState}) =>
+        useUserState({
+          idleCodes,
+          agentId,
+          cc: mockCC,
+          currentState,
+          customState: null,
+          lastStateChangeTimestamp: 1740744111287,
+          lastIdleCodeChangeTimestamp: undefined,
+          logger,
+          onStateChange,
+        }),
+      {initialProps: {currentState: '0', lastStateChangeTimestamp: 1740744111287}}
+    );
+
+    act(() => {
+      store.setCurrentState('2'); // Simulate the store state change
+      rerender({currentState: '2', lastStateChangeTimestamp: 1740748111287});
+    });
+
+    await waitFor(() => {
+      expect(mockCC.setAgentState).toHaveBeenCalledWith({
+        state: 'Available',
+        auxCodeId: '2',
+        agentId,
+        lastStateChangeReason: 'Available',
+      });
+      expect(store.lastStateChangeTimestamp).toEqual(resolvingValue.data.lastStateChangeTimestamp);
+    });
+  });
+
+  it('should set idle status if name does not match: Available', async () => {
+    const resolvingValue = {data: {auxCodeId: '1', lastStateChangeTimestamp: 1740748111287}};
+    mockCC.setAgentState.mockResolvedValueOnce(resolvingValue);
+    const {rerender} = renderHook(
+      ({currentState}) =>
+        useUserState({
+          idleCodes,
+          agentId,
+          cc: mockCC,
+          currentState,
+          customState: null,
+          lastStateChangeTimestamp: 1740744111287,
+          lastIdleCodeChangeTimestamp: undefined,
+          logger,
+          onStateChange,
+        }),
+      {initialProps: {currentState: '0'}}
+    );
+
+    act(() => {
+      store.setCurrentState('1'); // Simulate the store state change
+      rerender({currentState: '1'});
+    });
+
+    await waitFor(() => {
+      expect(mockCC.setAgentState).toHaveBeenCalledWith({
+        state: 'Idle',
+        auxCodeId: '1',
+        agentId,
+        lastStateChangeReason: 'Idle Code 1',
+      });
+      expect(store.lastStateChangeTimestamp).toEqual(resolvingValue.data.lastStateChangeTimestamp);
+    });
+  });
+
+  it('should handle errors from setAgentState and revert state', async () => {
     mockCC.setAgentState.mockRejectedValueOnce(new Error('Error setting agent status'));
-    const {result} = renderHook(() =>
-      useUserState({idleCodes, agentId, cc: mockCC, currentState: '0', lastStateChangeTimestamp: new Date()})
+    const {rerender} = renderHook(
+      ({currentState}) =>
+        useUserState({
+          idleCodes,
+          agentId,
+          cc: mockCC,
+          currentState,
+          customState: null,
+          lastStateChangeTimestamp: new Date(),
+          lastIdleCodeChangeTimestamp: undefined,
+          logger,
+          onStateChange,
+        }),
+      {initialProps: {currentState: '0'}}
     );
 
-    await act(async () => {
-      await result.current.setAgentStatus(idleCodes[1]);
+    act(() => {
+      store.setCurrentState('2'); // Simulate the store state change
+      rerender({currentState: '2'});
     });
 
     await waitFor(() => {
-      expect(result.current.errorMessage).toBe('Error: Error setting agent status');
+      expect(logger.error).toHaveBeenCalledWith('Error setting agent state: Error: Error setting agent status', {
+        module: 'useUserState',
+        method: 'updateAgentState',
+      });
     });
+  });
+
+  it('should handle stopIdleCodeTimer event and set lastIdleStateChangeElapsedTime to -1', () => {
+    const {result} = renderHook(() =>
+      useUserState({
+        idleCodes,
+        agentId,
+        cc: mockCC,
+        currentState: '0',
+        customState: null,
+        logger,
+        onStateChange,
+        lastStateChangeTimestamp: new Date().getTime(),
+        lastIdleCodeChangeTimestamp: new Date().getTime(),
+      })
+    );
+
+    act(() => {
+      workerMock.onmessage({data: {type: 'stopIdleCodeTimer'}});
+    });
+
+    expect(result.current.lastIdleStateChangeElapsedTime).toBe(-1);
+  });
+
+  it('should post resetIdleCode message if lastIdleCodeChangeTimestamp is different from lastStateChangeTimestamp', () => {
+    const lastStateChangeTimestamp = new Date().getTime();
+    const lastIdleCodeChangeTimestamp = lastStateChangeTimestamp - 1000; // 1 second earlier
+
+    renderHook(() =>
+      useUserState({
+        idleCodes,
+        agentId,
+        cc: mockCC,
+        currentState: '0',
+        customState: null,
+        logger,
+        onStateChange,
+        lastStateChangeTimestamp,
+        lastIdleCodeChangeTimestamp,
+      })
+    );
+
+    expect(workerMock.postMessage).toHaveBeenCalledWith({
+      type: 'resetIdleCode',
+      startTime: lastIdleCodeChangeTimestamp,
+    });
+  });
+
+  it('should not call onStateChange if not available', () => {
+    const customState = {developerName: 'Custom State'};
+    renderHook(() =>
+      useUserState({
+        idleCodes,
+        agentId,
+        cc: mockCC,
+        currentState: '0',
+        customState,
+        lastStateChangeTimestamp: new Date().getTime(),
+        lastIdleCodeChangeTimestamp: undefined,
+        logger,
+        onStateChange: undefined,
+      })
+    );
+
+    expect(onStateChange).not.toHaveBeenCalled();
+  });
+
+  it('should call onStateChange with customState if provided', () => {
+    const customState = {developerName: 'Custom State'};
+    renderHook(() =>
+      useUserState({
+        idleCodes,
+        agentId,
+        cc: mockCC,
+        currentState: '0',
+        customState,
+        lastStateChangeTimestamp: new Date(),
+        lastIdleCodeChangeTimestamp: undefined,
+        logger,
+        onStateChange,
+      })
+    );
+
+    expect(onStateChange).toHaveBeenCalledWith(customState);
+  });
+
+  it('should call onStateChange with matching idleCode when currentState changes', () => {
+    renderHook(() =>
+      useUserState({
+        idleCodes,
+        agentId,
+        cc: mockCC,
+        currentState: '1',
+        customState: null,
+        lastStateChangeTimestamp: new Date(),
+        lastIdleCodeChangeTimestamp: undefined,
+        logger,
+        onStateChange,
+      })
+    );
+
+    expect(onStateChange).toHaveBeenCalledWith(idleCodes[0]);
   });
 });